// projects/ais-uploader/src/lib/models/uploader-item.ts
export enum DocumentFileType {
  DOCX = 'application/vnd.openxmlformats-officedocument.wordprocessingml.document',
  DOC = 'application/msword',
  PDF = 'application/pdf',
  XLSX = 'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet',
  XLS = 'application/vnd.ms-excel',
  PPTX = 'application/vnd.openxmlformats-officedocument.presentationml.presentation',
  PPT = 'application/vnd.ms-powerpoint',
  TXT = 'text/plain',
  MP4 = 'video/mp4',
  AVI = 'video/avi',
  MPEG = 'video/mpeg',
  FLV = 'video/x-flv',
  WEBM = 'video/webm',
  OGG = 'video/ogg',
  JPEG = 'image/jpeg',
  JPG = 'image/jpeg',
  PNG = 'image/png',
  TIFF = 'image/tiff',
}